Move Inline editor into seperate component
